+++ tests/system_tests/test_hw_cpx_ap_i.py
@@ -569,43 +569,6 @@ def test_4iol_sdas(test_cpxap):
     m.write_module_parameter("Port Mode", "DEACTIVATED", sdas_channel)
 
 
-def test_4iol_sdas_full(test_cpxap):
-    m = test_cpxap.modules[4]
-    sdas_channel = 0
-
-    m.write_module_parameter("Port Mode", "IOL_AUTOSTART", sdas_channel)
-
-    time.sleep(0.05)
-
-    # example SDAS-MHS on port 0
-    param = m.read_fieldbus_parameters()
-
-    while param[sdas_channel]["Port status information"] != "OPERATE":
-        param = m.read_fieldbus_parameters()
-
-    sdas_data = m.read_channel(sdas_channel, full_size=True)
-    assert len(sdas_data) == 8
-
-    sdas_data = sdas_data[:2]  # only two bytes relevant
-
-    process_data = int.from_bytes(sdas_data, byteorder="big")
-
-    ssc1 = bool(process_data & 0x1)
-    ssc2 = bool(process_data & 0x2)
-    ssc3 = bool(process_data & 0x4)
-    ssc4 = bool(process_data & 0x8)
-    pdv = (process_data & 0xFFF0) >> 4
-
-    assert 0 <= pdv <= 4095
-
-    assert m[sdas_channel] == m.read_channel(sdas_channel)
-
-    m.write_module_parameter("Port Mode", "DEACTIVATED", sdas_channel)
-
-    while param[sdas_channel]["Port status information"] != "DEACTIVATED":
-        param = m.read_fieldbus_parameters()
-
-
 def test_4iol_sdas_read_isdu_string_as_raw(test_cpxap):
     m = test_cpxap.modules[4]
     sdas_channel = 0
@@ -1078,13 +1041,13 @@ def test_4iol_ethrottle(test_cpxap):
         data = module.read_channel(channel)
         # register order is [msb, ... , ... , lsb]
         # this unpack is not good, just for testing I can use legacy code
-        data = struct.unpack(">Q", data)[0]
+        data_int = struct.unpack(">Q", data)[0]
         process_input_data = {
-            "Actual Position": data & 0xFFFFFFFFFFFFFFF0,
-            "Homing Valid": bool(data & 8),
-            "Motion Complete": bool(data & 4),
-            "Proximity Switch": bool(data & 2),
-            "Reduced Speed": bool(data & 1),
+            "Actual Position": data_int & 0xFFFFFFFFFFFFFFF0,
+            "Homing Valid": bool(data_int & 8),
+            "Motion Complete": bool(data_int & 4),
+            "Proximity Switch": bool(data_int & 2),
+            "Reduced Speed": bool(data_int & 1),
         }
         return process_input_data